requirements.txt
matplotlib
numpy
python-chess
rtpt
sklearn
joblib
zarr
numcodecs
pytorch